chore: separate packages for subcommands
