+++ cmd/dump/dump.go
@@ -1,4 +1,4 @@
-package cmd
+package dump
 
 import (
 	"fmt"
@@ -6,6 +6,7 @@ import (
 
 	"gopkg.in/yaml.v3"
 
+	root "github.com/koh-sh/codebuild-multirunner/cmd"
 	"github.com/spf13/cobra"
 )
 
@@ -14,17 +15,17 @@ var dumpCmd = &cobra.Command{
 	Use:   "dump",
 	Short: "dump config for running CodeBuild projects",
 	Run: func(cmd *cobra.Command, args []string) {
-		bc := readConfigFile(configfile)
+		bc := root.ReadConfigFile(root.Configfile)
 		fmt.Println(dumpConfig(bc))
 	},
 }
 
 func init() {
-	rootCmd.AddCommand(dumpCmd)
+	root.RootCmd.AddCommand(dumpCmd)
 }
 
 // dump read config with environment variables inserted
-func dumpConfig(bc BuildConfig) string {
+func dumpConfig(bc root.BuildConfig) string {
 	d, err := yaml.Marshal(&bc)
 	if err != nil {
 		log.Fatal(err)
