+++ cmd/root.go
@@ -5,10 +5,14 @@ import (
 	"fmt"
 	"log"
 	"os"
+	"testing"
 
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs"
+	cwltypes "github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs/types"
 	"github.com/aws/aws-sdk-go-v2/service/codebuild"
+	"github.com/aws/aws-sdk-go-v2/service/codebuild/types"
+	"github.com/aws/smithy-go/middleware"
 	"github.com/spf13/cobra"
 	"gopkg.in/yaml.v3"
 )
@@ -146,10 +150,91 @@ type Build struct {
 	TimeoutInMinutesOverride         int                               `yaml:"timeoutInMinutesOverride,omitempty"`
 }
 
-var configfile string
+type StartBuildMockAPI func(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error)
+
+func (m StartBuildMockAPI) StartBuild(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error) {
+	return m(ctx, params, optFns...)
+}
+
+func (m StartBuildMockAPI) BatchGetBuilds(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error) {
+	return nil, nil
+}
+
+// return mock function for StartBuild
+func ReturnStartBuildMockAPI(build *types.Build, err error) func(t *testing.T) CodeBuildAPI {
+	mock := func(t *testing.T) CodeBuildAPI {
+		return StartBuildMockAPI(func(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error) {
+			t.Helper()
+			if params.ProjectName == nil {
+				t.Fatal("ProjectName is necessary")
+			}
+			return &codebuild.StartBuildOutput{
+				Build:          build,
+				ResultMetadata: middleware.Metadata{},
+			}, err
+		})
+	}
+	return mock
+}
+
+// mock api for BatchGetBuilds
+type BatchGetBuildsMockAPI func(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error)
+
+func (m BatchGetBuildsMockAPI) StartBuild(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error) {
+	return nil, nil
+}
+
+func (m BatchGetBuildsMockAPI) BatchGetBuilds(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error) {
+	return m(ctx, params, optFns...)
+}
+
+// return mock function for BatchgetBuilds
+func ReturnBatchGetBuildsMockAPI(builds []types.Build) func(t *testing.T) CodeBuildAPI {
+	mock := func(t *testing.T) CodeBuildAPI {
+		return BatchGetBuildsMockAPI(func(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error) {
+			t.Helper()
+			if len(params.Ids) == 0 {
+				t.Fatal("Ids must have at least one")
+			}
+			return &codebuild.BatchGetBuildsOutput{
+				Builds:         builds,
+				BuildsNotFound: []string{},
+				ResultMetadata: middleware.Metadata{},
+			}, nil
+		})
+	}
+	return mock
+}
+
+// mock api for GetLogEvents
+type GetLogEventsMockAPI func(ctx context.Context, params *cloudwatchlogs.GetLogEventsInput, optFns ...func(*cloudwatchlogs.Options)) (*cloudwatchlogs.GetLogEventsOutput, error)
+
+func (m GetLogEventsMockAPI) GetLogEvents(ctx context.Context, params *cloudwatchlogs.GetLogEventsInput, optFns ...func(*cloudwatchlogs.Options)) (*cloudwatchlogs.GetLogEventsOutput, error) {
+	return m(ctx, params, optFns...)
+}
+
+// return mock function for GetLogEvents
+func ReturnGetLogEventsMockAPI(events []cwltypes.OutputLogEvent) func(t *testing.T) CWLGetLogEventsAPI {
+	mock := func(t *testing.T) CWLGetLogEventsAPI {
+		return GetLogEventsMockAPI(func(ctx context.Context, params *cloudwatchlogs.GetLogEventsInput, optFns ...func(*cloudwatchlogs.Options)) (*cloudwatchlogs.GetLogEventsOutput, error) {
+			t.Helper()
+			if *params.LogGroupName == "" || *params.LogStreamName == "" {
+				t.Fatal("you must supply a logGroupName and logStreamName")
+			}
+			return &cloudwatchlogs.GetLogEventsOutput{
+				Events:            events,
+				NextBackwardToken: nil,
+				NextForwardToken:  nil,
+			}, nil
+		})
+	}
+	return mock
+}
+
+var Configfile string
 
 func SetVersionInfo(version, commit, date string) {
-	rootCmd.Version = fmt.Sprintf("%s (Built on %s from Git SHA %s)", version, date, commit)
+	RootCmd.Version = fmt.Sprintf("%s (Built on %s from Git SHA %s)", version, date, commit)
 }
 
 // interface for AWS CodeBuild API mock
@@ -182,7 +267,7 @@ func NewCloudWatchLogsAPI() (CWLGetLogEventsAPI, error) {
 }
 
 // read yaml config file for builds definition
-func readConfigFile(filepath string) BuildConfig {
+func ReadConfigFile(filepath string) BuildConfig {
 	bc := BuildConfig{}
 	b, err := os.ReadFile(filepath)
 	if err != nil {
@@ -197,7 +282,7 @@ func readConfigFile(filepath string) BuildConfig {
 }
 
 // rootCmd represents the base command when called without any subcommands
-var rootCmd = &cobra.Command{
+var RootCmd = &cobra.Command{
 	Use:   "codebuild-multirunner",
 	Short: "This is a simple CLI tool to \"Start build with overrides\" multiple AWS CodeBuild Projects at once.",
 	Long: `This is a simple CLI tool to "Start build with overrides" multiple AWS CodeBuild Projects at once.
@@ -207,12 +292,12 @@ This command will read YAML based config file and run multiple CodeBuild project
 }
 
 func Execute() {
-	err := rootCmd.Execute()
+	err := RootCmd.Execute()
 	if err != nil {
 		os.Exit(1)
 	}
 }
 
 func init() {
-	rootCmd.PersistentFlags().StringVar(&configfile, "config", "./config.yaml", "file path for config file.")
+	RootCmd.PersistentFlags().StringVar(&Configfile, "config", "./config.yaml", "file path for config file.")
 }
