+++ cmd/root_test.go
@@ -34,7 +34,7 @@ func Test_readConfigFile(t *testing.T) {
 	t.Setenv("TEST_ENV", "testproject3") // setting environment variable for test case 2
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := readConfigFile(tt.args.filepath); !reflect.DeepEqual(got, tt.want) {
+			if got := ReadConfigFile(tt.args.filepath); !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("readConfigFile() = %v, want %v", got, tt.want)
 			}
 		})
