--- cmd/run_test.go
@@ -1,212 +1,0 @@
-package cmd
-
-import (
-	"context"
-	"errors"
-	"reflect"
-	"testing"
-
-	"github.com/aws/aws-sdk-go-v2/service/codebuild"
-	"github.com/aws/aws-sdk-go-v2/service/codebuild/types"
-	"github.com/aws/smithy-go/middleware"
-	"github.com/fatih/color"
-)
-
-// TODO: get tidy for duplicated mocks
-// mock api for StartBuild
-type StartBuildMockAPI func(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error)
-
-func (m StartBuildMockAPI) StartBuild(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error) {
-	return m(ctx, params, optFns...)
-}
-
-func (m StartBuildMockAPI) BatchGetBuilds(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error) {
-	return nil, nil
-}
-
-// return mock function for StartBuild
-func ReturnStartBuildMockAPI(build *types.Build, err error) func(t *testing.T) CodeBuildAPI {
-	mock := func(t *testing.T) CodeBuildAPI {
-		return StartBuildMockAPI(func(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error) {
-			t.Helper()
-			if params.ProjectName == nil {
-				t.Fatal("ProjectName is necessary")
-			}
-			return &codebuild.StartBuildOutput{
-				Build:          build,
-				ResultMetadata: middleware.Metadata{},
-			}, err
-		})
-	}
-	return mock
-}
-
-// mock api for BatchGetBuilds
-type BatchGetBuildsMockAPI func(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error)
-
-func (m BatchGetBuildsMockAPI) StartBuild(ctx context.Context, params *codebuild.StartBuildInput, optFns ...func(*codebuild.Options)) (*codebuild.StartBuildOutput, error) {
-	return nil, nil
-}
-
-func (m BatchGetBuildsMockAPI) BatchGetBuilds(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error) {
-	return m(ctx, params, optFns...)
-}
-
-// return mock function for BatchgetBuilds
-func ReturnBatchGetBuildsMockAPI(builds []types.Build) func(t *testing.T) CodeBuildAPI {
-	mock := func(t *testing.T) CodeBuildAPI {
-		return BatchGetBuildsMockAPI(func(ctx context.Context, params *codebuild.BatchGetBuildsInput, optFns ...func(*codebuild.Options)) (*codebuild.BatchGetBuildsOutput, error) {
-			t.Helper()
-			if len(params.Ids) == 0 {
-				t.Fatal("Ids must have at least one")
-			}
-			return &codebuild.BatchGetBuildsOutput{
-				Builds:         builds,
-				BuildsNotFound: []string{},
-				ResultMetadata: middleware.Metadata{},
-			}, nil
-		})
-	}
-	return mock
-}
-
-func Test_convertBuildConfigToStartBuildInput(t *testing.T) {
-	type args struct {
-		build Build
-	}
-	tests := []struct {
-		name string
-		args args
-		want codebuild.StartBuildInput
-	}{
-		{name: "basic",
-			args: args{Build{}},
-			want: codebuild.StartBuildInput{},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := convertBuildConfigToStartBuildInput(tt.args.build); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("convertBuildConfigToStartBuildInput() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_coloredString(t *testing.T) {
-	type args struct {
-		status string
-	}
-	tests := []struct {
-		name string
-		args args
-		want string
-	}{
-		{name: "SUCCEEDED",
-			args: args{status: "SUCCEEDED"},
-			want: color.GreenString("SUCCEEDED")},
-		{name: "IN_PROGRESS",
-			args: args{status: "IN_PROGRESS"},
-			want: color.BlueString("IN_PROGRESS")},
-		{name: "FAILED",
-			args: args{status: "FAILED"},
-			want: color.RedString("FAILED")},
-		{name: "TIMED_OUT",
-			args: args{status: "TIMED_OUT"},
-			want: color.RedString("TIMED_OUT")},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := coloredString(tt.args.status); got != tt.want {
-				t.Errorf("coloredString() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_buildStatusCheck(t *testing.T) {
-	var id1 = "project:12345678"
-	var id2 = "project2:87654321"
-	var ids = []string{id1, id2}
-
-	type args struct {
-		client func(t *testing.T) CodeBuildAPI
-		ids    []string
-	}
-	tests := []struct {
-		name  string
-		args  args
-		want  []string
-		want2 bool
-	}{
-		{name: "all builds ended",
-			args:  args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "SUCCEEDED", Id: &id2}}), ids: ids},
-			want:  []string{},
-			want2: false,
-		},
-		{name: "one builds in progress",
-			args:  args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "IN_PROGRESS", Id: &id2}}), ids: ids},
-			want:  []string{id2},
-			want2: false,
-		},
-		{name: "one of builds failed",
-			args:  args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "FAILED", Id: &id2}}), ids: ids},
-			want:  []string{},
-			want2: true,
-		},
-		{name: "one of builds timeout",
-			args:  args{client: ReturnBatchGetBuildsMockAPI([]types.Build{{BuildStatus: "SUCCEEDED", Id: &id1}, {BuildStatus: "TIMED_OUT", Id: &id2}}), ids: ids},
-			want:  []string{},
-			want2: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, got2 := buildStatusCheck(tt.args.client(t), tt.args.ids)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("buildStatusCheck() = %v, want %v", got, tt.want)
-			}
-			if !reflect.DeepEqual(got2, tt.want2) {
-				t.Errorf("buildStatusCheck() = %v, want %v", got2, tt.want2)
-			}
-		})
-	}
-}
-
-func Test_runCodeBuild(t *testing.T) {
-	var project = "project"
-	var id = "project:12345"
-	type args struct {
-		client func(t *testing.T) CodeBuildAPI
-		input  codebuild.StartBuildInput
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    string
-		wantErr bool
-	}{
-		{name: "success to start",
-			args:    args{client: ReturnStartBuildMockAPI(&types.Build{Id: &id}, nil), input: codebuild.StartBuildInput{ProjectName: &project}},
-			want:    id,
-			wantErr: false,
-		},
-		{name: "fail to start",
-			args:    args{client: ReturnStartBuildMockAPI(&types.Build{Id: &id}, errors.New("fail to run")), input: codebuild.StartBuildInput{ProjectName: &project}},
-			want:    "",
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := runCodeBuild(tt.args.client(t), tt.args.input)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("runCodeBuild() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("runCodeBuild() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
