+++ main.go
@@ -1,6 +1,11 @@
 package main
 
-import "github.com/koh-sh/codebuild-multirunner/cmd"
+import (
+	"github.com/koh-sh/codebuild-multirunner/cmd"
+	_ "github.com/koh-sh/codebuild-multirunner/cmd/cwlog"
+	_ "github.com/koh-sh/codebuild-multirunner/cmd/dump"
+	_ "github.com/koh-sh/codebuild-multirunner/cmd/run"
+)
 
 var (
 	version = "dev"
